+++ tools/tests/test_template_parser.py
@@ -7,8 +7,6 @@ import unittest
 try:
     from tools.lib.template_parser import (
         TemplateParserException,
-        get_tag_info,
-        html_tag_tree,
         is_django_block_tag,
         tokenize,
         validate,
@@ -228,60 +226,3 @@ class ParserTest(unittest.TestCase):
         token = tokenize(tag)[0]
         self.assertEqual(token.kind, 'django_end')
         self.assertEqual(token.tag, 'if')
-
-    def test_get_tag_info(self):
-        # type: () -> None
-        html = '''
-            <p id="test" class="test1 test2">foo</p>
-        '''
-
-        start_tag, end_tag = tokenize(html)
-
-        start_tag_info = get_tag_info(start_tag)
-        end_tag_info = get_tag_info(end_tag)
-
-        self.assertEqual(start_tag_info.text(), 'p.test1.test2#test')
-        self.assertEqual(end_tag_info.text(), 'p')
-
-    def test_html_tag_tree(self):
-        # type: () -> None
-        html = '''
-            <!-- test -->
-            <!DOCTYPE html>
-            <html>
-            <!-- test -->
-            <head>
-                <title>Test</title>
-                <meta charset="utf-8" />
-                <link rel="stylesheet" href="style.css" />
-            </head>
-            <body>
-                <p>Hello<br />world!</p>
-                <p>Goodbye<!-- test -->world!</p>
-            </body>
-            </html>
-            <!-- test -->
-        '''
-
-        tree = html_tag_tree(html)
-
-        self.assertEqual(tree.children[0].token.kind, 'html_start')
-        self.assertEqual(tree.children[0].token.tag, 'html')
-
-        self.assertEqual(tree.children[0].children[0].token.kind, 'html_start')
-        self.assertEqual(tree.children[0].children[0].token.tag, 'head')
-
-        self.assertEqual(tree.children[0].children[0].children[0].token.kind, 'html_start')
-        self.assertEqual(tree.children[0].children[0].children[0].token.tag, 'title')
-
-        self.assertEqual(tree.children[0].children[1].token.kind, 'html_start')
-        self.assertEqual(tree.children[0].children[1].token.tag, 'body')
-
-        self.assertEqual(tree.children[0].children[1].children[0].token.kind, 'html_start')
-        self.assertEqual(tree.children[0].children[1].children[0].token.tag, 'p')
-
-        self.assertEqual(tree.children[0].children[1].children[0].children[0].token.kind, 'html_singleton')
-        self.assertEqual(tree.children[0].children[1].children[0].children[0].token.tag, 'br')
-
-        self.assertEqual(tree.children[0].children[1].children[1].token.kind, 'html_start')
-        self.assertEqual(tree.children[0].children[1].children[1].token.tag, 'p')
